requirements.txt
# AI Background Channel Studio - Backend Dependencies

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # ASGI server for FastAPI
python-multipart>=0.0.6  # For form data handling

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy>=2.0.23  # ORM for database operations
psycopg2-binary>=2.9.9  # PostgreSQL adapter
alembic>=1.12.1  # Database migrations (optional, for future)

# ============================================================================
# DATA VALIDATION & SERIALIZATION
# ============================================================================
pydantic>=2.5.0  # Data validation
pydantic-settings>=2.1.0  # Settings management

# ============================================================================
# ENVIRONMENT & CONFIGURATION
# ============================================================================
python-dotenv>=1.0.0  # Load environment variables from .env

# ============================================================================
# AI & LLM
# ============================================================================
openai>=1.3.0  # OpenAI API client for prompt and metadata generation

# ============================================================================
# HTTP REQUESTS
# ============================================================================
requests>=2.31.0  # HTTP library for API calls
httpx>=0.25.0  # Async HTTP client (used by FastAPI)

# ============================================================================
# MEDIA PROCESSING
# ============================================================================
pillow>=10.1.0  # Image manipulation (for dummy visual provider)
pydub>=0.25.1  # Audio manipulation
ffmpeg-python>=0.2.0  # Python wrapper for FFmpeg

# ============================================================================
# BACKGROUND JOBS (Optional for MVP, using simple polling)
# ============================================================================
# redis>=5.0.1  # Redis client for job queue
# rq>=1.15.1  # Redis Queue for background jobs

# ============================================================================
# TESTING
# ============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.1  # For async tests
httpx>=0.25.0  # For testing FastAPI endpoints

# ============================================================================
# DEVELOPMENT
# ============================================================================
black>=23.11.0  # Code formatter
flake8>=6.1.0  # Linting
mypy>=1.7.0  # Type checking
